refactor(ale_claw): fold image_sanitization into inference/, drop utils/
image_sanitization is provider-payload prep (resize/transcode so screenshots
fit provider image limits) — it belongs with the code that builds and sends
the request (unified_loop, _message_shapes), not in a 1-file utils/ package.
Move into inference/, remove the empty utils/ package. Behavior-preserving:
git mv + import-path updates (.utils -> .inference). 928/3. README updated.
