+++ ale_run/agents/ale_claw/harness/inference/image_sanitization.py
